requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==4.2.11
django-debug-toolbar==4.3.0
django-redis==5.4.0
djangorestframework==3.15.1
hiredis==2.3.2
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.11.0
tzdata==2024.1
